+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/deepseek
 
+## 2.0.0-beta.45
+
+### Patch Changes
+
+- Updated dependencies [3ed5519]
+  - @ai-sdk/provider-utils@4.0.0-beta.42
+
+## 2.0.0-beta.44
+
+### Patch Changes
+
+- 1bd7d32: feat: tool-specific strict mode
+- Updated dependencies [1bd7d32]
+  - @ai-sdk/provider-utils@4.0.0-beta.41
+  - @ai-sdk/provider@3.0.0-beta.23
+
+## 2.0.0-beta.43
+
+### Patch Changes
+
+- 4d04f43: feat(deepseek): rewrite DeepSeek provider
+
 ## 2.0.0-beta.42
 
 ### Patch Changes