+++ packages/huggingface/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/huggingface
 
+## 1.0.0-beta.39
+
+### Patch Changes
+
+- Updated dependencies [3ed5519]
+  - @ai-sdk/provider-utils@4.0.0-beta.42
+  - @ai-sdk/openai-compatible@2.0.0-beta.43
+
+## 1.0.0-beta.38
+
+### Patch Changes
+
+- Updated dependencies [1bd7d32]
+  - @ai-sdk/openai-compatible@2.0.0-beta.42
+  - @ai-sdk/provider-utils@4.0.0-beta.41
+  - @ai-sdk/provider@3.0.0-beta.23
+
 ## 1.0.0-beta.37
 
 ### Patch Changes