+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/openai-compatible
 
+## 2.0.0-beta.43
+
+### Patch Changes
+
+- Updated dependencies [3ed5519]
+  - @ai-sdk/provider-utils@4.0.0-beta.42
+
+## 2.0.0-beta.42
+
+### Patch Changes
+
+- 1bd7d32: feat: tool-specific strict mode
+- Updated dependencies [1bd7d32]
+  - @ai-sdk/provider-utils@4.0.0-beta.41
+  - @ai-sdk/provider@3.0.0-beta.23
+
 ## 2.0.0-beta.41
 
 ### Patch Changes