+++ packages/perplexity/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/perplexity
 
+## 3.0.0-beta.44
+
+### Patch Changes
+
+- Updated dependencies [3ed5519]
+  - @ai-sdk/provider-utils@4.0.0-beta.42
+
+## 3.0.0-beta.43
+
+### Patch Changes
+
+- Updated dependencies [1bd7d32]
+  - @ai-sdk/provider-utils@4.0.0-beta.41
+  - @ai-sdk/provider@3.0.0-beta.23
+
 ## 3.0.0-beta.42
 
 ### Patch Changes