+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/valibot
 
+## 2.0.0-beta.42
+
+### Patch Changes
+
+- Updated dependencies [3ed5519]
+  - @ai-sdk/provider-utils@4.0.0-beta.42
+
+## 2.0.0-beta.41
+
+### Patch Changes
+
+- Updated dependencies [1bd7d32]
+  - @ai-sdk/provider-utils@4.0.0-beta.41
+
 ## 2.0.0-beta.40
 
 ### Patch Changes